coze2openai/requirements.txt
Flask==2.1.1
flask-cors==3.0.10
requests==2.28.1
python-dotenv==0.20.0
Werkzeug==2.1.2
